pyproject.toml
[tool.poetry]
name = "akshit-mountain-sinai-medical-center-designing-api"
version = "0.1.0"
description = ""
authors = ["jk"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
djangorestframework = "^3.13.0"
django-simple-history = "^3.0.0"
django-cors-headers = "^3.10.1"
psycopg2-binary = "^2.9.2"
django-extensions = "^3.1.5"
drf-yasg = "^1.20.0"
djangorestframework-simplejwt = "^5.0.0"
phonenumbers = "^8.12.39"
pyotp = "^2.6.0"
gunicorn = "^20.1.0"
stripe = "^2.64.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
isort = "^5.10.1"
flake8 = "^4.0.1"
tox = "^3.24.4"
pygraphviz = "^1.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
